Linting, move env to community
