+++ environments/community/physical_space_stl/README.md
@@ -33,4 +33,3 @@ $ sudo apt-get install libglfw3-dev libgles2-mesa-dev libnvidia-gl-570-server
 
 Generated run: https://wandb.ai/csxl/atropos-environments_hack0/runs/dlexyg5r
 Training run (ran out of memory): https://wandb.ai/csxl/grpo-physical-trainer/runs/t61am7gu
-
