+++ environments/community/physical_space_stl/llm_label.py
@@ -1,10 +1,15 @@
-import os
 import json
-import trimesh
+import os
+
 import torch
+import trimesh
 from PIL import Image
-from transformers import BlipProcessor, BlipForConditionalGeneration
-from transformers import AutoModelForCausalLM, AutoTokenizer
+from transformers import (
+    AutoModelForCausalLM,
+    AutoTokenizer,
+    BlipForConditionalGeneration,
+    BlipProcessor,
+)
 
 device = "cuda" if torch.cuda.is_available() else "cpu"
 
@@ -14,12 +19,17 @@ LABEL_FILE = "dataset/labels.json"
 
 # Load BLIP for image captioning
 blip_processor = BlipProcessor.from_pretrained("Salesforce/blip-image-captioning-base")
-blip_model = BlipForConditionalGeneration.from_pretrained("Salesforce/blip-image-captioning-base").to(device)
+blip_model = BlipForConditionalGeneration.from_pretrained(
+    "Salesforce/blip-image-captioning-base"
+).to(device)
 
 # Load Mistral or other small LLM
 llm_model_name = "mistralai/Mistral-7B-Instruct-v0.2"
 llm_tokenizer = AutoTokenizer.from_pretrained(llm_model_name)
-llm_model = AutoModelForCausalLM.from_pretrained(llm_model_name, torch_dtype=torch.float16, device_map="auto")
+llm_model = AutoModelForCausalLM.from_pretrained(
+    llm_model_name, torch_dtype=torch.float16, device_map="auto"
+)
+
 
 def extract_trimesh_features(mesh):
     return {
@@ -32,12 +42,14 @@ def extract_trimesh_features(mesh):
         "euler_number": mesh.euler_number,
     }
 
+
 def caption_image(image_path):
     raw_image = Image.open(image_path).convert("RGB")
     inputs = blip_processor(raw_image, return_tensors="pt").to(device)
     out = blip_model.generate(**inputs)
     caption = blip_processor.decode(out[0], skip_special_tokens=True)
     return caption
+
 
 def generate_label_with_llm(features, caption):
     prompt = f"""You are a 3D object classifier.
@@ -55,6 +67,7 @@ Label:"""
     outputs = llm_model.generate(**inputs, max_new_tokens=10, do_sample=False)
     output_text = llm_tokenizer.decode(outputs[0], skip_special_tokens=True)
     return output_text.split("Label:")[-1].strip()
+
 
 def main():
     labels = {}
@@ -86,6 +99,6 @@ def main():
 
     print(f"\nSaved labels to {LABEL_FILE}")
 
+
 if __name__ == "__main__":
     main()
-
