+++ environments/community/physical_space_stl/render_stl.py
@@ -1,17 +1,18 @@
-import bpy
-import sys
 import math
 import os
+import sys
+
+import bpy
 
 # Get args after --
 argv = sys.argv
-argv = argv[argv.index("--") + 1:]  # args after --
+argv = argv[argv.index("--") + 1 :]  # args after --
 
 input_stl = argv[0]
 output_dir = argv[1]
 
 # Clear existing objects
-bpy.ops.object.select_all(action='SELECT')
+bpy.ops.object.select_all(action="SELECT")
 bpy.ops.object.delete(use_global=False)
 
 # Import STL
@@ -19,11 +20,11 @@ bpy.ops.import_mesh.stl(filepath=input_stl)
 obj = bpy.context.selected_objects[0]
 
 # Center the object at origin
-bpy.ops.object.origin_set(type='ORIGIN_CENTER_OF_MASS', center='MEDIAN')
+bpy.ops.object.origin_set(type="ORIGIN_CENTER_OF_MASS", center="MEDIAN")
 obj.location = (0, 0, 0)
 
 # Add Sun light
-sun_light_data = bpy.data.lights.new(name="SunLight", type='SUN')
+sun_light_data = bpy.data.lights.new(name="SunLight", type="SUN")
 sun_light_object = bpy.data.objects.new(name="SunLight", object_data=sun_light_data)
 sun_light_object.location = (10, 10, 10)
 bpy.context.collection.objects.link(sun_light_object)
@@ -52,10 +53,9 @@ for i, angle in enumerate(angles):
 
     # Point camera to object center (0,0,0)
     direction = -cam_obj.location
-    rot_quat = direction.to_track_quat('-Z', 'Y')
+    rot_quat = direction.to_track_quat("-Z", "Y")
     cam_obj.rotation_euler = rot_quat.to_euler()
 
     # Render
     bpy.context.scene.render.filepath = os.path.join(output_dir, f"render_{i}.png")
     bpy.ops.render.render(write_still=True)
-
