+++ environments/community/physical_space_stl/test_rendered_sphere_views.png
